[ControllerRegistration] Prefactor: extract defaulting into dedicated file
