+++ pkg/apis/core/v1beta1/defaults_test.go
@@ -29,125 +29,6 @@ import (
 )
 
 var _ = Describe("Defaults", func() {
-	Describe("#SetDefaults_ControllerResource", func() {
-		It("should default the primary field", func() {
-			resource := ControllerResource{}
-
-			SetDefaults_ControllerResource(&resource)
-
-			Expect(resource.Primary).To(PointTo(BeTrue()))
-		})
-
-		It("should not default the primary field", func() {
-			resource := ControllerResource{Primary: pointer.Bool(false)}
-			resourceCopy := resource.DeepCopy()
-
-			SetDefaults_ControllerResource(&resource)
-
-			Expect(resource.Primary).To(Equal(resourceCopy.Primary))
-		})
-
-		const kindExtension = "Extension"
-		It("should default the globallyEnabled field when kind is Extension", func() {
-			resource := ControllerResource{Kind: kindExtension}
-			SetDefaults_ControllerResource(&resource)
-
-			Expect(resource.GloballyEnabled).To(Equal(pointer.Bool(false)))
-		})
-
-		It("should not default the globallyEnabled field when kind is Extension and globallyEnabled is already set", func() {
-			resource := ControllerResource{Kind: kindExtension, GloballyEnabled: pointer.Bool(true)}
-			SetDefaults_ControllerResource(&resource)
-
-			Expect(resource.GloballyEnabled).To(Equal(pointer.Bool(true)))
-		})
-
-		It("should not default the globallyEnabled field when kind is not Extension", func() {
-			resource := ControllerResource{Kind: "not extension"}
-			SetDefaults_ControllerResource(&resource)
-
-			Expect(resource.GloballyEnabled).To(BeNil())
-		})
-
-		It("should default the reconcile timeout when kind is Extension", func() {
-			resource := ControllerResource{Kind: kindExtension}
-			SetDefaults_ControllerResource(&resource)
-
-			Expect(resource.ReconcileTimeout).To(Equal(&metav1.Duration{Duration: time.Minute * 3}))
-		})
-
-		It("should not default the reconcile timeout when kind is Extension and timeout is already set", func() {
-			resource := ControllerResource{Kind: kindExtension, ReconcileTimeout: &metav1.Duration{Duration: time.Second * 62}}
-			SetDefaults_ControllerResource(&resource)
-
-			Expect(resource.ReconcileTimeout).To(Equal(&metav1.Duration{Duration: time.Second * 62}))
-		})
-
-		It("should not default the reconcile timeout when kind is not Extension", func() {
-			resource := ControllerResource{Kind: "not extension"}
-			SetDefaults_ControllerResource(&resource)
-
-			Expect(resource.ReconcileTimeout).To(BeNil())
-		})
-
-		It("should default the lifecycle strategy field when kind is Extension", func() {
-			resource := ControllerResource{Kind: kindExtension}
-			SetDefaults_ControllerResource(&resource)
-
-			Expect(resource.Lifecycle).ToNot(BeNil())
-			Expect(*resource.Lifecycle.Reconcile).To(Equal(AfterKubeAPIServer))
-			Expect(*resource.Lifecycle.Delete).To(Equal(BeforeKubeAPIServer))
-			Expect(*resource.Lifecycle.Migrate).To(Equal(BeforeKubeAPIServer))
-		})
-
-		It("should default the lifecycle strategy field when kind is not Extension", func() {
-			resource := ControllerResource{Kind: "not an extension"}
-			SetDefaults_ControllerResource(&resource)
-
-			Expect(resource.Lifecycle).To(BeNil())
-		})
-
-		It("should default only the missing lifecycle strategy fields when kind is Extension", func() {
-			before := BeforeKubeAPIServer
-			resource := ControllerResource{
-				Kind: kindExtension,
-				Lifecycle: &ControllerResourceLifecycle{
-					Reconcile: &before,
-				},
-			}
-			SetDefaults_ControllerResource(&resource)
-
-			Expect(resource.Lifecycle).ToNot(BeNil())
-			Expect(*resource.Lifecycle.Reconcile).To(Equal(BeforeKubeAPIServer))
-			Expect(*resource.Lifecycle.Delete).To(Equal(BeforeKubeAPIServer))
-			Expect(*resource.Lifecycle.Migrate).To(Equal(BeforeKubeAPIServer))
-		})
-	})
-
-	Describe("#SetDefaults_ControllerDeployment", func() {
-		var (
-			ondemand = ControllerDeploymentPolicyOnDemand
-			always   = ControllerDeploymentPolicyAlways
-		)
-
-		It("should default the policy field", func() {
-			deployment := ControllerRegistrationDeployment{}
-
-			SetDefaults_ControllerRegistrationDeployment(&deployment)
-
-			Expect(deployment.Policy).To(PointTo(Equal(ondemand)))
-		})
-
-		It("should not default the policy field", func() {
-			deployment := ControllerRegistrationDeployment{Policy: &always}
-			deploymentCopy := deployment.DeepCopy()
-
-			SetDefaults_ControllerRegistrationDeployment(&deployment)
-
-			Expect(deployment.Policy).To(Equal(deploymentCopy.Policy))
-		})
-	})
-
 	Describe("#SetObjectDefaults_Seed", func() {
 		var obj *Seed
